Move sierra debug info to separate module
No business logic changes besides making local functions private

commit-id:5312705a
