+++ crates/cairo-lang-sierra-generator/src/lib.rs
@@ -6,6 +6,7 @@ mod ap_tracking;
 mod block_generator;
 pub mod canonical_id_replacer;
 pub mod db;
+pub mod debug_info;
 pub mod executables;
 mod expr_generator_context;
 mod extra_sierra_info;
@@ -21,9 +22,6 @@ pub mod program_generator;
 pub mod replace_ids;
 mod resolve_labels;
 mod specialization_context;
-pub mod statements_code_locations;
-pub mod statements_functions;
-pub mod statements_locations;
 mod store_variables;
 #[cfg(any(feature = "testing", test))]
 pub mod test_utils;
